# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Net Zone CRM API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  agent: 'agent',
} as const;

export interface User {
  id: string;
  username: string;
  name: string;
  role: UserRole;
  isActive: boolean;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}

export interface Customer {
  id: string;
  name: string;
  mobile: string;
  /** @nullable */
  alternateNumber?: string | null;
  email: string;
  /** @nullable */
  company?: string | null;
  address: string;
  /** @nullable */
  city?: string | null;
  category: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  followUpDate?: string | null;
  /** @nullable */
  totalCalls?: number | null;
  /** @nullable */
  lastCallAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CustomerInput {
  name: string;
  mobile: string;
  alternateNumber?: string;
  email?: string;
  company?: string;
  address?: string;
  city?: string;
  category?: string;
  notes?: string;
  followUpDate?: string;
}

export interface CustomerUpdate {
  name?: string;
  mobile?: string;
  alternateNumber?: string;
  email?: string;
  company?: string;
  address?: string;
  city?: string;
  category?: string;
  notes?: string;
  followUpDate?: string;
}

export type CallType = typeof CallType[keyof typeof CallType];


export const CallType = {
  incoming: 'incoming',
  outgoing: 'outgoing',
  missed: 'missed',
} as const;

export interface Call {
  id: string;
  customerId: string;
  /** @nullable */
  customerName?: string | null;
  agentId: string;
  /** @nullable */
  agentName?: string | null;
  type: CallType;
  duration: number;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  remark?: string | null;
  /** @nullable */
  followUpDate?: string | null;
  /** @nullable */
  reminderDate?: string | null;
  /** @nullable */
  deviceId?: string | null;
  createdAt: string;
}

export type CallInputType = typeof CallInputType[keyof typeof CallInputType];


export const CallInputType = {
  incoming: 'incoming',
  outgoing: 'outgoing',
  missed: 'missed',
} as const;

export interface CallInput {
  customerId: string;
  type: CallInputType;
  duration: number;
  phoneNumber?: string;
  category?: string;
  remark?: string;
  followUpDate?: string;
  reminderDate?: string;
  deviceId?: string;
}

export interface CallUpdate {
  type?: string;
  duration?: number;
  remark?: string;
  category?: string;
  followUpDate?: string;
  reminderDate?: string;
}

export interface BulkSyncInput {
  calls: CallInput[];
}

export interface BulkSyncResult {
  synced: number;
  failed: number;
}

export interface Remark {
  id: string;
  customerId: string;
  agentId: string;
  text: string;
  createdAt: string;
}

export interface RemarkInput {
  customerId: string;
  text: string;
}

export interface RemarkUpdate {
  text: string;
}

export type ReminderStatus = typeof ReminderStatus[keyof typeof ReminderStatus];


export const ReminderStatus = {
  pending: 'pending',
  completed: 'completed',
  overdue: 'overdue',
} as const;

export type ReminderPriority = typeof ReminderPriority[keyof typeof ReminderPriority];


export const ReminderPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Reminder {
  id: string;
  customerId: string;
  agentId: string;
  title: string;
  dateTime: string;
  /** @nullable */
  notes?: string | null;
  status: ReminderStatus;
  priority: ReminderPriority;
  createdAt: string;
}

export type ReminderInputPriority = typeof ReminderInputPriority[keyof typeof ReminderInputPriority];


export const ReminderInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface ReminderInput {
  customerId: string;
  title: string;
  dateTime: string;
  notes?: string;
  priority: ReminderInputPriority;
}

export type ReminderUpdateStatus = typeof ReminderUpdateStatus[keyof typeof ReminderUpdateStatus];


export const ReminderUpdateStatus = {
  pending: 'pending',
  completed: 'completed',
  overdue: 'overdue',
} as const;

export interface ReminderUpdate {
  status?: ReminderUpdateStatus;
  title?: string;
  dateTime?: string;
  notes?: string;
}

export interface Category {
  id: string;
  name: string;
  color: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  color: string;
  description?: string;
}

export interface CategoryUpdate {
  name?: string;
  color?: string;
  description?: string;
}

export type AgentRole = typeof AgentRole[keyof typeof AgentRole];


export const AgentRole = {
  admin: 'admin',
  agent: 'agent',
} as const;

export interface Agent {
  id: string;
  username: string;
  name: string;
  role: AgentRole;
  isActive: boolean;
  /** @nullable */
  totalCalls?: number | null;
  /** @nullable */
  todayCalls?: number | null;
  createdAt: string;
}

export type AgentInputRole = typeof AgentInputRole[keyof typeof AgentInputRole];


export const AgentInputRole = {
  admin: 'admin',
  agent: 'agent',
} as const;

export interface AgentInput {
  username: string;
  name: string;
  password: string;
  role: AgentInputRole;
}

export type AgentUpdateRole = typeof AgentUpdateRole[keyof typeof AgentUpdateRole];


export const AgentUpdateRole = {
  admin: 'admin',
  agent: 'agent',
} as const;

export interface AgentUpdate {
  name?: string;
  role?: AgentUpdateRole;
  isActive?: boolean;
  password?: string;
}

export interface DashboardStats {
  todayCalls: number;
  missedCalls: number;
  followUps: number;
  reminders: number;
  totalCustomers: number;
  incomingCalls: number;
  outgoingCalls: number;
}

export interface AgentPerformance {
  agentId: string;
  agentName: string;
  totalCalls: number;
  incomingCalls: number;
  outgoingCalls: number;
  missedCalls: number;
  avgDuration: number;
}

export interface RepeatCaller {
  customerId: string;
  customerName: string;
  mobile: string;
  callCount: number;
  /** @nullable */
  category?: string | null;
}

export interface DailyCallStat {
  date: string;
  total: number;
  incoming: number;
  outgoing: number;
  missed: number;
}

export interface ReportSummary {
  totalCalls: number;
  incomingCalls: number;
  outgoingCalls: number;
  missedCalls: number;
  avgDuration: number;
  byDay: DailyCallStat[];
}

export type ListCustomersParams = {
search?: string;
category?: string;
page?: number;
limit?: number;
};

export type ListCallsParams = {
period?: ListCallsPeriod;
agentId?: string;
type?: string;
customerId?: string;
page?: number;
limit?: number;
};

export type ListCallsPeriod = typeof ListCallsPeriod[keyof typeof ListCallsPeriod];


export const ListCallsPeriod = {
  today: 'today',
  yesterday: 'yesterday',
  week: 'week',
} as const;

export type ListRemarksParams = {
customerId?: string;
};

export type ListRemindersParams = {
customerId?: string;
status?: string;
};

export type GetAgentPerformanceParams = {
period?: GetAgentPerformancePeriod;
};

export type GetAgentPerformancePeriod = typeof GetAgentPerformancePeriod[keyof typeof GetAgentPerformancePeriod];


export const GetAgentPerformancePeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
} as const;

export type GetCallsReportParams = {
from?: string;
to?: string;
agentId?: string;
type?: string;
};

export type GetReportSummaryParams = {
from?: string;
to?: string;
agentId?: string;
};

export type GetCustomersReportParams = {
category?: string;
from?: string;
to?: string;
};