# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Net Zone CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── CALLS ───────────────────────────────────────────────────────────────────
...
  /remarks:
    get:
      operationId: listRemarks
      tags: [remarks]
      summary: List remarks, optionally by customer
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of remarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Remark"
    post:
      operationId: createRemark
      tags: [remarks]
      summary: Create a remark
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemarkInput"
      responses:
        "201":
          description: Remark created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Remark"
  /remarks/{id}:
    put:
      operationId: updateRemark
      tags: [remarks]
      summary: Update a remark
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemarkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Remark"
    delete:
      operationId: deleteRemark
      tags: [remarks]
      summary: Delete a remark
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── REMINDERS ───────────────────────────────────────────────────────────────
...
  /dashboard:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get today's stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/agent-performance:
    get:
      operationId: getAgentPerformance
      tags: [dashboard]
      summary: Agent performance stats
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month]
      responses:
        "200":
          description: Agent performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentPerformance"
...
  /dashboard/repeat-callers:
    get:
      operationId: getRepeatCallers
      tags: [dashboard]
      summary: Customers who called multiple times today
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Repeat callers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepeatCaller"

  # ── REPORTS ─────────────────────────────────────────────────────────────────
...
  /reports/customers:
    get:
      operationId: getCustomersReport
      tags: [reports]
      summary: Customers report with call counts
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers with call stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"

  # ── AGENTS ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      required: [id, name, mobile, email, address, category, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        mobile:
          type: string
        alternateNumber:
          type: ["string", "null"]
        email:
          type: string
        company:
          type: ["string", "null"]
        address:
          type: string
        city:
          type: ["string", "null"]
        category:
          type: string
        notes:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        totalCalls:
          type: ["integer", "null"]
        lastCallAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CustomerInput:
      type: object
      required: [name, mobile]
      properties:
        name:
          type: string
        mobile:
          type: string
        alternateNumber:
          type: string
        email:
          type: string
        company:
          type: string
        address:
          type: string
        city:
          type: string
        category:
          type: string
        notes:
          type: string
        followUpDate:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        alternateNumber:
          type: string
        email:
          type: string
        company:
          type: string
        address:
          type: string
        city:
          type: string
        category:
          type: string
        notes:
          type: string
        followUpDate:
          type: string

    # ── CALLS ─────────────────────────────────────────────────────────────────
...
    Remark:
      type: object
      required: [id, customerId, agentId, text, createdAt]
      properties:
        id:
          type: string
        customerId:
          type: string
        agentId:
          type: string
        text:
          type: string
        createdAt:
          type: string
    RemarkInput:
      type: object
      required: [customerId, text]
      properties:
        customerId:
          type: string
        text:
          type: string
    RemarkUpdate:
      type: object
      required: [text]
      properties:
        text:
          type: string

    # ── REMINDERS ─────────────────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      required: [todayCalls, missedCalls, followUps, reminders, totalCustomers, incomingCalls, outgoingCalls]
      properties:
        todayCalls:
          type: integer
        missedCalls:
          type: integer
        followUps:
          type: integer
        reminders:
          type: integer
        totalCustomers:
          type: integer
        incomingCalls:
          type: integer
        outgoingCalls:
          type: integer
    AgentPerformance:
      type: object
      required: [agentId, agentName, totalCalls, incomingCalls, outgoingCalls, missedCalls, avgDuration]
      properties:
        agentId:
          type: string
        agentName:
          type: string
        totalCalls:
          type: integer
        incomingCalls:
          type: integer
        outgoingCalls:
          type: integer
        missedCalls:
          type: integer
        avgDuration:
          type: number
    RepeatCaller:
      type: object
      required: [customerId, customerName, mobile, callCount]
      properties:
        customerId:
          type: string
        customerName:
          type: string
        mobile:
          type: string
        callCount:
          type: integer
        category:
          type: ["string", "null"]

    # ── REPORTS ───────────────────────────────────────────────────────────────
...