# lib/api-zod/src/generated/types/agentPerformance.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Net Zone CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentPerformance {
  agentId: string;
  agentName: string;
  totalCalls: number;
  incomingCalls: number;
  outgoingCalls: number;
  missedCalls: number;
  avgDuration: number;
}